sync

+++ src/test/java/com/dalendev/finance/cryptobot/util/PriceUtilTest.java
@@ -12,11 +12,9 @@ public class PriceUtilTest {
 
     @Test
     public void adjust() throws Exception {
-
-        assertEquals(new Float(115.0f), PriceUtil.adjust(114.73999786f, 1.0f));
-        assertEquals(new Float(115.0f), PriceUtil.adjust(115.0f, 1.0f));
-        assertEquals(new Float(2.33f), PriceUtil.adjust(2.32999992f, 0.01f));
-
+        assertEquals(new Double(115.0f), PriceUtil.adjust(114.73999786d, 1.0d));
+        assertEquals(new Double(115.0f), PriceUtil.adjust(115.0d, 1.0d));
+        assertEquals(new Double(2.33f), PriceUtil.adjust(2.32999992d, 0.01d));
     }
 
 }
